+++ src/i18n/i18n.ts
@@ -1,5 +1,6 @@
 import i18n from 'i18next';
 import { initReactI18next } from 'react-i18next';
+import LanguageDetector from 'i18next-browser-languagedetector';
 import type { LanguageMetadata } from './types';
 import { en, enMetadata } from './locales/en';
 import { es, esMetadata } from './locales/es';
@@ -7,12 +8,20 @@ import { fr, frMetadata } from './locales/fr';
 import { de, deMetadata } from './locales/de';
 import { hi, hiMetadata } from './locales/hi';
 import { ja, jaMetadata } from './locales/ja';
-import { ko_KR, ko_KRMetadata } from './locales/ko_KR.ts';
+import { ko_KR, ko_KRMetadata } from './locales/ko_KR';
 import { pt_BR, pt_BRMetadata } from './locales/pt_BR';
 import { uk, ukMetadata } from './locales/uk';
 import { ru, ruMetadata } from './locales/ru';
 import { zh_CN, zh_CNMetadata } from './locales/zh_CN';
 import { zh_TW, zh_TWMetadata } from './locales/zh_TW';
+import { ne, neMetadata } from './locales/ne';
+import { mr, mrMetadata } from './locales/mr';
+import { tr, trMetadata } from './locales/tr';
+import { id_ID, id_IDMetadata } from './locales/id_ID';
+import { te, teMetadata } from './locales/te';
+import { bn, bnMetadata } from './locales/bn';
+import { gu, guMetadata } from './locales/gu';
+import { vi, viMetadata } from './locales/vi';
 
 export const languages: LanguageMetadata[] = [
     enMetadata,
@@ -27,6 +36,14 @@ export const languages: LanguageMetadata[] = [
     ruMetadata,
     zh_CNMetadata,
     zh_TWMetadata,
+    neMetadata,
+    mrMetadata,
+    trMetadata,
+    id_IDMetadata,
+    teMetadata,
+    bnMetadata,
+    guMetadata,
+    viMetadata,
 ];
 
 const resources = {
@@ -42,16 +59,25 @@ const resources = {
     ru,
     zh_CN,
     zh_TW,
+    ne,
+    mr,
+    tr,
+    id_ID,
+    te,
+    bn,
+    gu,
+    vi,
 };
 
-i18n.use(initReactI18next).init({
-    resources,
-    lng: enMetadata.code,
-    interpolation: {
-        escapeValue: false,
-    },
-    fallbackLng: enMetadata.code,
-    debug: false,
-});
+i18n.use(LanguageDetector)
+    .use(initReactI18next)
+    .init({
+        resources,
+        interpolation: {
+            escapeValue: false,
+        },
+        fallbackLng: enMetadata.code,
+        debug: false,
+    });
 
 export { i18n };