+++ src/i18n/locales/de.ts
@@ -28,7 +28,6 @@ export const de: LanguageTranslation = {
                 show_cardinality: 'Kardinalität anzeigen',
                 zoom_on_scroll: 'Zoom beim Scrollen',
                 theme: 'Stil',
-                change_language: 'Sprache',
                 show_dependencies: 'Abhängigkeiten anzeigen',
                 hide_dependencies: 'Abhängigkeiten ausblenden',
             },
@@ -77,6 +76,18 @@ export const de: LanguageTranslation = {
             dont_show_again: 'Nicht erneut anzeigen',
             change_schema: 'Schema ändern',
             none: 'Keine',
+        },
+
+        copy_to_clipboard_toast: {
+            unsupported: {
+                title: 'Kopieren fehlgeschlagen',
+                description: 'Zwischenablage nicht unterstützt',
+            },
+            failed: {
+                title: 'Kopieren fehlgeschlagen',
+                description:
+                    'Etwas ist schiefgelaufen. Bitte versuchen Sie es erneut.',
+            },
         },
 
         theme: {
@@ -345,6 +356,11 @@ export const de: LanguageTranslation = {
             format_json: 'JSON',
             cancel: 'Cancel',
             export: 'Export',
+            error: {
+                title: 'Error exporting diagram',
+                description:
+                    'Something went wrong. Need help? chartdb.io@gmail.com',
+            },
         },
         // TODO: Translate
         import_diagram_dialog: {
@@ -382,10 +398,15 @@ export const de: LanguageTranslation = {
         tool_tips: {
             double_click_to_edit: 'Doppelklicken zum Bearbeiten',
         },
+
+        language_select: {
+            change_language: 'Sprache',
+        },
     },
 };
 
 export const deMetadata: LanguageMetadata = {
-    name: 'Deutsch',
+    name: 'German',
+    nativeName: 'Deutsch',
     code: 'de',
 };