+++ src/i18n/locales/es.ts
@@ -28,7 +28,6 @@ export const es: LanguageTranslation = {
                 hide_sidebar: 'Ocultar Barra Lateral',
                 zoom_on_scroll: 'Zoom al Desplazarse',
                 theme: 'Tema',
-                change_language: 'Idioma',
                 show_dependencies: 'Mostrar dependencias',
                 hide_dependencies: 'Ocultar dependencias',
             },
@@ -68,6 +67,17 @@ export const es: LanguageTranslation = {
                 'Esta acción reorganizará todas las tablas en el diagrama. ¿Deseas continuar?',
             reorder: 'Reordenar',
             cancel: 'Cancelar',
+        },
+
+        copy_to_clipboard_toast: {
+            unsupported: {
+                title: 'Copia fallida',
+                description: 'Portapapeles no soportado',
+            },
+            failed: {
+                title: 'Copia fallida',
+                description: 'Algo salió mal. Por favor, inténtelo de nuevo.',
+            },
         },
 
         theme: {
@@ -345,6 +355,11 @@ export const es: LanguageTranslation = {
             format_json: 'JSON',
             cancel: 'Cancel',
             export: 'Export',
+            error: {
+                title: 'Error exporting diagram',
+                description:
+                    'Something went wrong. Need help? chartdb.io@gmail.com',
+            },
         },
         // TODO: Translate
         import_diagram_dialog: {
@@ -382,10 +397,15 @@ export const es: LanguageTranslation = {
         tool_tips: {
             double_click_to_edit: 'Doble clic para editar',
         },
+
+        language_select: {
+            change_language: 'Idioma',
+        },
     },
 };
 
 export const esMetadata: LanguageMetadata = {
-    name: 'Español',
+    name: 'Spanish',
+    nativeName: 'Español',
     code: 'es',
 };