+++ src/i18n/locales/hi.ts
@@ -28,7 +28,6 @@ export const hi: LanguageTranslation = {
                 show_cardinality: 'कार्डिनैलिटी दिखाएँ',
                 zoom_on_scroll: 'स्क्रॉल पर ज़ूम',
                 theme: 'थीम',
-                change_language: 'भाषा बदलें',
                 show_dependencies: 'निर्भरता दिखाएँ',
                 hide_dependencies: 'निर्भरता छिपाएँ',
             },
@@ -77,6 +76,17 @@ export const hi: LanguageTranslation = {
             dont_show_again: 'फिर से न दिखाएँ',
             change_schema: 'बदलें',
             none: 'कोई नहीं',
+        },
+
+        copy_to_clipboard_toast: {
+            unsupported: {
+                title: 'कॉपी असफल',
+                description: 'क्लिपबोर्ड समर्थित नहीं है',
+            },
+            failed: {
+                title: 'कॉपी असफल',
+                description: 'कुछ गलत हो गया। कृपया पुनः प्रयास करें।',
+            },
         },
 
         theme: {
@@ -347,6 +357,11 @@ export const hi: LanguageTranslation = {
             format_json: 'JSON',
             cancel: 'Cancel',
             export: 'Export',
+            error: {
+                title: 'Error exporting diagram',
+                description:
+                    'Something went wrong. Need help? chartdb.io@gmail.com',
+            },
         },
         // TODO: Translate
         import_diagram_dialog: {
@@ -384,10 +399,15 @@ export const hi: LanguageTranslation = {
         tool_tips: {
             double_click_to_edit: 'संपादित करने के लिए डबल-क्लिक करें',
         },
+
+        language_select: {
+            change_language: 'भाषा बदलें',
+        },
     },
 };
 
 export const hiMetadata: LanguageMetadata = {
     name: 'Hindi',
+    nativeName: 'हिन्दी',
     code: 'hi',
 };