+++ src/i18n/locales/ko_KR.ts
@@ -28,15 +28,13 @@ export const ko_KR: LanguageTranslation = {
                 show_cardinality: '카디널리티 보이기',
                 zoom_on_scroll: '스크롤 시 확대',
                 theme: '테마',
-                change_language: '언어/Language',
                 show_dependencies: '종속성 보이기',
                 hide_dependencies: '종속성 숨기기',
             },
-            // TODO: Translate
             share: {
-                share: 'Share',
-                export_diagram: 'Export Diagram',
-                import_diagram: 'Import Diagram',
+                share: '공유',
+                export_diagram: '다이어그램 내보내기',
+                import_diagram: '다이어그램 가져오기',
             },
             help: {
                 help: '도움말',
@@ -77,6 +75,17 @@ export const ko_KR: LanguageTranslation = {
             dont_show_again: '다시 보여주지 마세요',
             change_schema: '변경',
             none: '없음',
+        },
+
+        copy_to_clipboard_toast: {
+            unsupported: {
+                title: '복사 실패',
+                description: '클립보드가 지원되지 않습니다"',
+            },
+            failed: {
+                title: '복사 실패',
+                description: '문제가 발생했습니다. 다시 시도해주세요.',
+            },
         },
 
         theme: {
@@ -145,7 +154,7 @@ export const ko_KR: LanguageTranslation = {
                         change_schema: '스키마 변경',
                         add_field: '필드 추가',
                         add_index: '인덱스 추가',
-                        duplicate_table: 'Duplicate Table', // TODO: Translate
+                        duplicate_table: '테이블 복제',
                         delete_table: '테이블 삭제',
                     },
                 },
@@ -232,8 +241,7 @@ export const ko_KR: LanguageTranslation = {
 
             cancel: '취소',
             back: '뒤로가기',
-            // TODO: Translate
-            import_from_file: 'Import from File',
+            import_from_file: '파일에서 가져오기',
             empty_diagram: '빈 다이어그램으로 시작',
             continue: '계속',
             import: '가져오기',
@@ -336,24 +344,27 @@ export const ko_KR: LanguageTranslation = {
             close: '아직은 괜찮아요',
             confirm: '당연하죠!',
         },
-        // TODO: Translate
         export_diagram_dialog: {
-            title: 'Export Diagram',
-            description: 'Choose the format for export:',
+            title: '다이어그램 내보내기',
+            description: '내보낼 형식을 선택해주세요:',
             format_json: 'JSON',
-            cancel: 'Cancel',
-            export: 'Export',
-        },
-        // TODO: Translate
-        import_diagram_dialog: {
-            title: 'Import Diagram',
-            description: 'Paste the diagram JSON below:',
-            cancel: 'Cancel',
-            import: 'Import',
+            cancel: '취소',
+            export: '내보내기',
             error: {
-                title: 'Error importing diagram',
+                title: '다이어그램 내보내기 오류',
                 description:
-                    'The diagram JSON is invalid. Please check the JSON and try again. Need help? chartdb.io@gmail.com',
+                    '무언가 문제가 발생하였습니다. 도움이 필요하신 경우 chartdb.io@gmail.com으로 연락해주세요.',
+            },
+        },
+        import_diagram_dialog: {
+            title: '다이어그램 가져오기',
+            description: '아래에 다이어그램 JSON을 첨부해주세요:',
+            cancel: '취소',
+            import: '가져오기',
+            error: {
+                title: '다이어그램 가져오기 오류',
+                description:
+                    '다이어그램 JSON이 유효하지 않습니다. JSON이 올바른 형식인지 확인해주세요. 도움이 필요하신 경우 chartdb.io@gmail.com으로 연락해주세요.',
             },
         },
         relationship_type: {
@@ -370,20 +381,24 @@ export const ko_KR: LanguageTranslation = {
 
         table_node_context_menu: {
             edit_table: '테이블 수정',
-            duplicate_table: 'Duplicate Table', // TODO: Translate
+            duplicate_table: '테이블 복제',
             delete_table: '테이블 삭제',
         },
 
-        // TODO: Add translations
-        snap_to_grid_tooltip: 'Snap to Grid (Hold {{key}})',
+        snap_to_grid_tooltip: '그리드에 맞추기 ({{key}}를 누른채 유지)',
 
         tool_tips: {
             double_click_to_edit: '더블클릭하여 편집',
+        },
+
+        language_select: {
+            change_language: '언어',
         },
     },
 };
 
 export const ko_KRMetadata: LanguageMetadata = {
-    name: '한국어',
+    name: 'Korean',
+    nativeName: '한국어',
     code: 'ko_KR',
 };