+++ src/i18n/locales/pt_BR.ts
@@ -28,7 +28,6 @@ export const pt_BR: LanguageTranslation = {
                 show_cardinality: 'Mostrar Cardinalidade',
                 zoom_on_scroll: 'Zoom ao Rolar',
                 theme: 'Tema',
-                change_language: 'Idioma',
                 show_dependencies: 'Mostrar Dependências',
                 hide_dependencies: 'Ocultar Dependências',
             },
@@ -77,6 +76,17 @@ export const pt_BR: LanguageTranslation = {
             dont_show_again: 'Não mostrar novamente',
             change_schema: 'Alterar',
             none: 'nenhum',
+        },
+
+        copy_to_clipboard_toast: {
+            unsupported: {
+                title: 'Falha na cópia',
+                description: 'Área de transferência não suportada',
+            },
+            failed: {
+                title: 'Falha na cópia',
+                description: 'Algo deu errado. Por favor, tente novamente.',
+            },
         },
 
         theme: {
@@ -344,6 +354,11 @@ export const pt_BR: LanguageTranslation = {
             format_json: 'JSON',
             cancel: 'Cancel',
             export: 'Export',
+            error: {
+                title: 'Error exporting diagram',
+                description:
+                    'Something went wrong. Need help? chartdb.io@gmail.com',
+            },
         },
         // TODO: Translate
         import_diagram_dialog: {
@@ -381,10 +396,15 @@ export const pt_BR: LanguageTranslation = {
         tool_tips: {
             double_click_to_edit: 'Duplo clique para editar',
         },
+
+        language_select: {
+            change_language: 'Idioma',
+        },
     },
 };
 
 export const pt_BRMetadata: LanguageMetadata = {
     name: 'Portuguese',
+    nativeName: 'Português',
     code: 'pt_BR',
 };