+++ src/pages/editor-page/editor-desktop-layout.tsx
@@ -6,9 +6,9 @@ import {
 } from '@/components/resizable/resizable';
 import { SidePanel } from './side-panel/side-panel';
 import { Canvas } from './canvas/canvas';
-import { useBreakpoint } from '@/hooks/use-breakpoint';
 import { useLayout } from '@/hooks/use-layout';
 import type { Diagram } from '@/lib/domain/diagram';
+import { cn } from '@/lib/utils';
 
 export interface EditorDesktopLayoutProps {
     initialDiagram?: Diagram;
@@ -17,22 +17,21 @@ export const EditorDesktopLayout: React.FC<EditorDesktopLayoutProps> = ({
     initialDiagram,
 }) => {
     const { isSidePanelShowed } = useLayout();
-    const { isLg } = useBreakpoint('lg');
-    const { isXl } = useBreakpoint('xl');
 
     return (
         <ResizablePanelGroup direction="horizontal">
             <ResizablePanel
-                defaultSize={isXl ? 25 : isLg ? 35 : 50}
-                minSize={isXl ? 25 : isLg ? 35 : 50}
+                defaultSize={25}
+                minSize={25}
                 maxSize={isSidePanelShowed ? 99 : 0}
-                // eslint-disable-next-line
-                className="transition-[flex-grow] duration-200"
+                className={cn('transition-[flex-grow] duration-200', {
+                    'min-w-[350px]': isSidePanelShowed,
+                })}
             >
                 <SidePanel />
             </ResizablePanel>
-            <ResizableHandle />
-            <ResizablePanel defaultSize={isXl ? 75 : isLg ? 65 : 50}>
+            <ResizableHandle disabled={!isSidePanelShowed} />
+            <ResizablePanel defaultSize={75}>
                 <Canvas initialTables={initialDiagram?.tables ?? []} />
             </ResizablePanel>
         </ResizablePanelGroup>